requirements.txt
ffmpeg
scenedetect[opencv]
opencv-python
imagehash
Pillow
git+https://github.com/openai/whisper.git
sentence-transformers
ffmpeg-python
ollama
pyannote-audio
torch
torchvision
torchaudio
python-dotenv
tiktoken